Do not migrate unmapped pages
If a page in the source range is unmapped, we do not want to allocate a
destination page for it.

Previously, checking for MIGRATE_PFN_MIGRATE was enough to accomplish
this, but after commit 8315ada7f095 ("mm/migrate: allow migrate_vma() to
alloc new page on empty entry"), we now need an additional check.