migrate_pages: Do not force starting node 0
The test is initialized by moving all the allocated pages to a starting
node. Then the test migrates them to a destination node.

Now that our destination nodes are in ZONE_DEVICE, there is no need to
do the initialization. We are already guaranteed that all the allocated
pages are from a non-destination node.